pyproject.toml
[tool.poetry]
name = "leap"
version = "1.0.0"
description = "Multi-session Claude Code with message queueing and image support"
authors = ["Nevo Mashiach"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/nevo24/leap"
repository = "https://github.com/nevo24/leap"

[tool.poetry.dependencies]
python = "^3.11"
pexpect = "^4.8.0"
prompt-toolkit = "^3.0.0"
gnureadline = "^8.1.2"
wcwidth = "^0.6.0"
rich = "^14.3.2"
pyobjc-framework-cocoa = "^12.1"
pyobjc-framework-applicationservices = "^12.1"
pyobjc-framework-quartz = "^12.1"
pyte = "^0.8.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"

[tool.poetry.group.monitor.dependencies]
PyQt5 = "^5.15.0"
python-gitlab = "^8.0.0"
PyGithub = "^2.0"
py2app = "^0.28"

[tool.poetry.group.slack]
optional = true

[tool.poetry.group.slack.dependencies]
slack-bolt = "^1.18"
slack-sdk = "^3.27"

[tool.poetry.scripts]
claudel = "leap.main:main"
leap-server = "leap.server.server:main"
leap-client = "leap.client.client:main"
leap-monitor = "leap.monitor.app:main"
leap-slack = "leap.slack.bot:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
